+++ examples/triangle_glfw.py
@@ -5,7 +5,7 @@ The glfw library can be installed using ``pip install glfw``.
 
 import glfw
 from wgpu.gui.glfw import WgpuCanvas  # WgpuCanvas wraps a glfw window
-import wgpu.backend.rs  # noqa: F401, Select Rust backend
+import wgpu.backends.rs  # noqa: F401, Select Rust backend
 
 # Import the (async) function that we must call to run the visualization
 from triangle import main
