+++ examples/triangle_glfw_asyncio.py
@@ -8,7 +8,7 @@ import asyncio  # noqa: E402
 
 import glfw
 from wgpu.gui.glfw import WgpuCanvas  # WgpuCanvas wraps a glfw window
-import wgpu.backend.rs  # noqa: F401, Select Rust backend
+import wgpu.backends.rs  # noqa: F401, Select Rust backend
 
 # Import the (async) function that we must call to run the visualization
 from triangle import main_async
