+++ wgpu/__pyinstaller/test_wgpu.py
@@ -1,6 +1,6 @@
 def test_pyi_wgpu(pyi_builder):
     pyi_builder.test_source(
         """
-        import wgpu.backend.rs
+        import wgpu.backends.rs
     """
     )
