--- wgpu/backend/__init__.py
@@ -1,4 +1,0 @@
-"""
-The backend implementations of the wgpu API.
-You need to select one by importing it, e.g. ``import wgpu.backend.rs``.
-"""
